+++ src/index.js
@@ -1,8 +1,9 @@
 import Discord from 'discord.js';
 import fs from 'fs';
 import { LoadConfig, UpdateConfig } from './config';
-import TeamGen from './teamgenerator';
+import { FormattedTeams, GenerateTeams } from './teamgenerator';
 import { StartPresenceCycler, StopPresenceCycler, SetPresence, PresenceOff } from './presence';
+import { DeleteMessage, GetChannelUsers, GetVoiceChannel, SendMessage, GetRoleID } from './utils';
 
 const client = new Discord.Client();
 
@@ -17,128 +18,6 @@ LoadConfig().then((config) => {
       StartPresenceCycler(config.presencecycletime, client);
     }
   });
-
-  const TeamsToFields = (teams) => {
-    const fields = [];
-    
-    for (const key of Object.keys(teams)) {
-      let players = '';
-
-      teams[key].players.forEach((player) => {
-        players = `${players}${player}\n`;
-      });
-
-      const field = {
-        name: key,
-        value: players
-      }
-
-      fields.push(field);
-    }
-
-    return fields;
-  };
-
-  const TeamsToString = (teams) => {
-    let message = '';
-
-    for (const key of Object.keys(teams)) {
-      let players = '';
-
-      teams[key].players.forEach((player) => {
-        players = `${players}${player}\n`;
-      });
-
-      message = `${message}**${key}**\n${players}\n\n`;
-    }
-
-    message = `${message}*Now go forth..and multiply!*`;
-
-    return message;
-  }
-
-  const GetTeamsEmbed = (teams) => {
-    return {
-      embed: {
-        color: 3447003,
-        author: {
-          name: client.user.username,
-          icon_url: client.user.avatarURL
-        },
-        title: "The teams are...",
-        fields: TeamsToFields(teams),
-        timestamp: new Date(),
-        footer: {
-          icon_url: client.user.avatarURL,
-          text: "© GFX Mofa phuka"
-        }
-      }
-    }
-  };
-
-  const GetVoiceChannel = (message) => {
-    let voiceChannel;
-    let playerFoundInVoice = false;
-    message.guild.channels.forEach(channel => {
-      if (channel.type === 'voice') {
-        channel.members.forEach(member => {
-          if (member.id == message.author.id) {
-            playerFoundInVoice = true;
-            voiceChannel = channel;
-          }
-        })
-      }
-    });
-
-    if (!playerFoundInVoice) {
-      throw new Error('Did not find player in voice');
-    }
-
-    return voiceChannel;
-  };
-
-  const GetVoicePlayers = (message) => {
-    
-    const PlayerList = [];
-    try {
-      const voiceChannel = GetVoiceChannel(message);
-      voiceChannel.members.forEach(member => {
-        PlayerList.push(member.user.toString());
-      });
-
-      return PlayerList;
-    } catch(e) {
-      throw new Error('No voice channel');
-    }
-  };
-
-  const CleanUpMessage = (message) => {
-    message.delete(config.autocleanup*1000).catch((error) => {
-      if (config.debuginchat) {
-        message.channel.send(`I can't delete a message: ${error}`);
-      }
-    });
-  }
-
-  const SendToChannel = (originalMessage, data, autoDeleteable = false) => {
-    const responseMessage = originalMessage.channel.send(data);
-    responseMessage.then((message) => {
-      if (config.autocleanup > 0 && autoDeleteable) {
-        CleanUpMessage(message);
-        CleanUpMessage(originalMessage);
-      }
-    });
-  };
-
-  const GetRoleID = (guild, roleName) => {
-    const role = guild.roles.find("name", roleName);
-
-    if (role) {
-      return role.id;
-    } else {
-      return false;
-    }
-  };
 
   const SetCMDCooldown = () => {
     CommandCooldown = true;
@@ -171,7 +50,7 @@ LoadConfig().then((config) => {
 
         if (args.length < 1) {
           try {
-            playerList = GetVoicePlayers(message)
+            playerList = GetChannelUsers(GetVoiceChannel(message));
           } catch (e) {
             message.channel.send('You need to be in a voice channel, idiot...');
             message.channel.send('Or you can send a list of names (seperated by a space) i.e. !teams 2 Player1 Player2 Player3');
@@ -184,12 +63,12 @@ LoadConfig().then((config) => {
 
 
 
-        const Teams = TeamGen.GenerateTeams(
-          numberOfTeams || 2,
-          playerList
+        const Teams = FormattedTeams(
+          GenerateTeams(numberOfTeams || 2, playerList),
+          'string'
         );
 
-        SendToChannel(message, TeamsToString(Teams), true);
+        SendMessage(message, Teams, (config.autocleanup > 0));
       }
 
       if (command === 'dumpconfig' && isAdmin) {
@@ -201,13 +80,13 @@ LoadConfig().then((config) => {
         SetCMDCooldown();
         if (args.length == 2) {
           UpdateConfig(args[0], args[1]);
-          SendToChannel(message, `Updated ${args[0]} to ${args[1]}.`, true);
+          SendMessage(message, `Updated ${args[0]} to ${args[1]}.`, (config.autocleanup > 0));
         }
       }
 
       if (command === 'loot' || command === 'graveh') {
         SetCMDCooldown();
-        SendToChannel(message, `*sprints and loots ${args.join(' ') || 'everything'} before ${message.author.toString()} can get there*`);
+        SendMessage(message, `*sprints and loots ${args.join(' ') || 'everything'} before ${message.author.toString()} can get there*`);
       }
 
       if(command === 'presence' && isAdmin) {
@@ -239,14 +118,14 @@ LoadConfig().then((config) => {
           }
 
           if (config.autocleanup > 0) {
-            CleanUpMessage(message);
+            DeleteMessage(message, config.autocleanup);
           }
         }
       }
 
       if (command === 'ping') {
         SetCMDCooldown();
-        SendToChannel(message, 'pong!', true);
+        SendMessage(message, 'pong!', (config.autocleanup > 0));
       }
