+++ src/utils/index.js
@@ -1,0 +1,13 @@
+import DeleteMessage from './delete-message';
+import GetChannelUsers from './get-channel-users';
+import GetVoiceChannel from './get-voice-cannel';
+import SendMessage from './send-message';
+import GetRoleID from './get-role-id';
+
+export default {
+    DeleteMessage,
+    GetChannelUsers,
+    GetVoiceChannel,
+    SendMessage,
+    GetRoleID
+}
